+++ CHANGELOG.md
@@ -1,7 +1,17 @@
+## 1.1.0
+
+* Bumped `compileSdk` to 34.
+* Updated `compileSdk` and `targetSdkVersion` of example app to 34.
+* Flutter SDK version will be bumped to make it easier to maintain the plugin.
+* Update a dependency to the latest release.
+* Added a mechanism for confirming the delivery of a message.
+* ClientTokens can now be deleted if they are not used for a long time. And in this case, they will be updated.
+* Added logging of the service in the debug mode (FlutterPushedMessaging.getLog()).
+
 ## 1.0.4
 
-- Minor fixes
+* Minor fixes
 
 ## 1.0.1
 
-- Initial release.
+* Initial release.